remove bash
